eventis/requirements.txt
django
djangorestframework
requests
mysqlclient
rest_authtoken